MEDIUM: resolv: separate resolving and dns message exchange layers on send path
This patch attempts to split the code between the resolving and
the DNS message exchange layers on the send path of a request

Previous code sadly re-send all pending resolution reaching an eagain.
This patch also disable this and the sent message is considered
lost, the upper layer should handle the retry.